fixed fucking everything

+++ src/main/java/com/momchil/TU4ALL/Constants.java
@@ -1,8 +1,8 @@
 package com.momchil.TU4ALL;
 
 public final class Constants {
-    public static final long FRIEND_STATUS_ADD = 0;
-    public static final long FRIEND_STATUS_REQUESTED = 1;
-    public static final long FRIEND_STATUS_ACCEPTED = 2;
-    public static final long FRIEND_STATUS_BLOCKED = 3;
+    public static final int FRIEND_STATUS_ADD = 0;
+    public static final int FRIEND_STATUS_REQUESTED = 1;
+    public static final int FRIEND_STATUS_ACCEPTED = 2;
+    public static final int FRIEND_STATUS_BLOCKED = 3;
 }
